jwt parsing update (#55)
